+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 Language Improvements:
 
+- fix(cpp) improve parsing issues with templates (#2752) [Josh Goebel][]
+- enh(cpp) add support for `enum (struct|class)` and `union` (#2752) [Josh Goebel][]
 - fix(js/ts) Fix nesting of `{}` inside template literals SUBST expression (#2748) [Josh Goebel][]
 - enh(js/ts) Highlight class methods as functions (#2727) [Josh Goebel][]
 - fix(js/ts) `constructor` is now highlighted as a function title (not keyword) (#2727) [Josh Goebel][]